s3: implement upload of folders
`mkdir()` is no-op for S3, we get creation of parent folders for free.
The only thing we need is an hack in `stat()`: we always have to assume
that if the HEAD request fails then it's a virtual folder.
That's because KIO calls stat() on parent folders without a trailing
slash in the path.