Fix compilation with aws-sdk-cpp built with custom memory management
If the aws-sdk-cpp is built with custom memory management,
Aws::String is no longer equivalent to a std::string.

So we need to explicitly use Aws::String and manually convert to/from QString where necessary.

Note: for bucket names it's enough to use Latin-1, while for
keys/prefixes we need to use UTF8.